refactor(teams): split recipe_bridge.rs into focused submodules [OPE-407]
Decompose the monolithic recipe_bridge.rs (1062 LOC) into a module
directory with clear boundaries:

- conversion.rs: profile_to_recipe, recipe_to_profile, settings_to_retry_config,
  and parameter type parsing/formatting helpers
- extensions.rs: ext_ref_to_config, config_to_ext_ref for bidirectional
  ExtensionRef/ExtensionConfig mapping
- tests.rs: all 14 existing tests preserved
- mod.rs: module coordinator with public re-exports

All 274 crate tests pass (257 unit + 17 integration), zero clippy
warnings, clean formatting.
